Move CostsListScreen to a separate directory
