+++ src/screens/CostsListScreen/index.ts
@@ -1,0 +1,1 @@
+export { default, CostsListScreenNavOptions } from './CostsListScreen';
